Redesign
In this commit, we introduce a new design for this project: Compared to the previous design, which was based on standalone binaries written in C that implement certain well-defined tasks and are composable by pipes, i.e. following the UNIX philosophy, we replaced it by an implementation in Rust based. Perspectively, the code presented in lowpass::osc is intended as a general framework for the numerical tasks connected to synthesis, and shall be composable in sequencers and mixers by using traits.